% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CMASE_obs.R
\name{compute_CMASE_obs}
\alias{compute_CMASE_obs}
\title{Compute Circular Mean Average Squared Error (Observational)}
\usage{
compute_CMASE_obs(theta, theta_hat)
}
\arguments{
\item{theta}{Numeric or circular vector of observed responses (in radians).}

\item{theta_hat}{Numeric or circular vector of predicted responses (in radians).}
}
\value{
A scalar: the mean of the squared circular errors, i.e., CMASE_obs.
}
\description{
Computes the empirical circular mean squared error between observed responses
and predicted values. This corresponds to the observational version of CMASE
(Circular Mean Average Squared Error) as defined in the article, and is the
circular analogue of the usual MSE.
}
\examples{
theta <- circular::circular(runif(100, 0, 2 * pi))
theta_hat <- circular::circular(runif(100, 0, 2 * pi))
compute_CMASE_obs(theta, theta_hat)

}
