% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R2_circular.R
\name{compute_R2_circular}
\alias{compute_R2_circular}
\title{Compute Circular Coefficient of Determination R^2}
\usage{
compute_R2_circular(theta, theta_hat)
}
\arguments{
\item{theta}{Vector of observed circular responses (in radians).}

\item{theta_hat}{Vector of predicted circular responses (in radians).}
}
\value{
A scalar representing the circular coefficient of determination.
}
\description{
Computes the circular analogue of the coefficient of determination,
based on the cosine distance between observed and predicted angular responses.
}
\details{
The formula used is:
\deqn{
  R^2_{\text{circ}} = 1 - \frac{\sum (1 - \cos(\theta_i - \hat{\theta}_i))}{\sum (1 - \cos(\theta_i - \bar{\theta}))}
}{
  R^2_circ = 1 - SSE / SST
}
where \eqn{\bar{\theta}} is the circular mean of the observed angles.
}
\examples{
set.seed(123)
theta <- circular::circular(runif(100, 0, 2 * pi))
theta_hat <- circular::circular(runif(100, 0, 2 * pi))
compute_R2_circular(theta, theta_hat)

}
