% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simultaneous_bands.R
\name{compute_simultaneous_bands}
\alias{compute_simultaneous_bands}
\title{Compute Simultaneous Bootstrap Confidence Bands (Nadaraya-Watson)}
\usage{
compute_simultaneous_bands(...)
}
\arguments{
\item{...}{Arguments passed to \code{compute_simultaneous_bands_general}.}
}
\value{
Same as \code{compute_simultaneous_bands_general()}.
}
\description{
Wrapper for \code{compute_simultaneous_bands_general()} using \code{nw_circular_predict}.
}
