% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simultaneous_bands_LL.R
\name{compute_simultaneous_bands_LL}
\alias{compute_simultaneous_bands_LL}
\title{Compute Simultaneous Bootstrap Confidence Bands (Local Linear)}
\usage{
compute_simultaneous_bands_LL(...)
}
\arguments{
\item{...}{Arguments passed to \code{compute_simultaneous_bands_general}.}
}
\value{
Same as \code{compute_simultaneous_bands_general()}.
}
\description{
Wrapper for \code{compute_simultaneous_bands_general()} using \code{ll_circular_predict}.
}
