% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_histogram.R
\name{plot_residuals_histogram}
\alias{plot_residuals_histogram}
\title{Plot Circular Histogram of Residuals}
\usage{
plot_residuals_histogram(
  resids,
  bins = 16,
  main = "Circular Histogram of Residuals"
)
}
\arguments{
\item{resids}{Numeric vector of residuals (in radians).}

\item{bins}{Integer. Number of bins in the circular histogram. Default is 16.}

\item{main}{Character. Title of the plot. Default is \code{"Circular Histogram of Residuals"}.}
}
\value{
A circular histogram plot created with \code{rose.diag()}.
}
\description{
Displays a circular histogram (rose diagram) of the residuals.
This is useful to visually assess the distribution of residuals
in circular regression models.
}
\examples{
resids <- circular::circular(rnorm(100, mean = 0, sd = 1), units = "radians")
plot_residuals_histogram(resids, bins = 12, main = "Example: Residuals")

}
