% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_violin_by_group_interactive.R
\name{plot_residuals_violin_by_group_interactive}
\alias{plot_residuals_violin_by_group_interactive}
\title{Violin Plot of Residuals by Group (Interactive)}
\usage{
plot_residuals_violin_by_group_interactive(
  resids,
  groups,
  main = "Violin Plot of Residuals by Group"
)
}
\arguments{
\item{resids}{Vector of residuals (in radians).}

\item{groups}{Factor or character vector of group labels.}

\item{main}{Title of the plot.}
}
\value{
A ggplot2 object.
}
\description{
Creates a violin plot of circular residuals grouped by a categorical variable.
}
\examples{
set.seed(123)
resids <- runif(100, -pi, pi)
groups <- sample(c("A", "B", "C"), 100, replace = TRUE)
plot_residuals_violin_by_group_interactive(resids, groups)

}
