% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_violin_by_group_pdf.R
\name{plot_residuals_violin_by_group_pdf}
\alias{plot_residuals_violin_by_group_pdf}
\title{Save Violin Plot of Residuals by Group to PDF}
\usage{
plot_residuals_violin_by_group_pdf(
  resids,
  groups,
  file = "violin_residuals.pdf",
  main = "Violin Plot of Residuals by Group"
)
}
\arguments{
\item{resids}{Vector of residuals (in radians).}

\item{groups}{Factor or character vector of group labels.}

\item{file}{Output PDF file name.}

\item{main}{Title of the plot.}
}
\value{
Invisibly returns the path to the saved PDF file.
}
\description{
Creates and saves a violin plot of circular residuals grouped by a categorical variable.
}
\examples{
set.seed(123)
resids <- runif(100, -pi, pi)
groups <- sample(c("A", "B", "C"), 100, replace = TRUE)
plot_file <- tempfile(fileext = ".pdf")
plot_residuals_violin_by_group_pdf(resids, groups, file = plot_file)

}
