% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_of_thumb_bandwidth.R
\name{rule_of_thumb_bandwidth}
\alias{rule_of_thumb_bandwidth}
\title{Rule of Thumb Bandwidth Selector for Circular Regression Estimation}
\usage{
rule_of_thumb_bandwidth(
  data,
  cont_cols = NULL,
  cat_cols = NULL,
  estimator = "NW",
  ch_NW = 1.06,
  clambda_NW = 1,
  ch_LL = 1.2,
  clambda_LL = 1,
  exponent_cont = 1/5,
  exponent_cat = 1/5,
  alpha = 1
)
}
\arguments{
\item{data}{Data frame containing the sample.}

\item{cont_cols}{Character vector of names of continuous predictors.}

\item{cat_cols}{Character vector of names of categorical predictors.}

\item{estimator}{Character string: either "NW" (default) or "LL".}

\item{ch_NW, clambda_NW}{Constants for NW selector (defaults: 1.06 and 1.0).}

\item{ch_LL, clambda_LL}{Constants for LL selector (defaults: 1.20 and 1.0).}

\item{exponent_cont}{Exponent for sample size in the rule for continuous predictors (default = 1/5).}

\item{exponent_cat}{Exponent for sample size in the rule for categorical predictors (default = 1/5).}

\item{alpha}{Exponent for the number of levels of each categorical predictor (default = 1).}
}
\value{
A list with numeric vectors \code{h} (for continuous predictors) and \code{lambda} (for categorical ones).
}
\description{
Provides heuristic values for bandwidths in circular regression models
with continuous and categorical predictors, adapted to either
Nadaraya--Watson (NW) or local linear (LL) estimators.
}
\details{
The rule applies classic scaling laws based on asymptotic MSE minimization.
Different constants can be used depending on the estimator.
}
\examples{
data <- data.frame(x = runif(100), z = sample(letters[1:3], 100, replace = TRUE))
rule_of_thumb_bandwidth(data, cont_cols = "x", cat_cols = "z", estimator = "LL")

}
