% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_residuals_uniformity.R
\name{test_residuals_uniformity}
\alias{test_residuals_uniformity}
\title{Test Uniformity of Circular Residuals}
\usage{
test_residuals_uniformity(resids)
}
\arguments{
\item{resids}{Numeric vector of residuals (in radians).}
}
\value{
A list with components:
\describe{
\item{rayleigh}{Result of the Rayleigh test.}
\item{kuiper}{Result of the Kuiper test.}
\item{watson}{Result of the Watson \eqn{U^2} test.}
}
}
\description{
Applies Rayleigh, Kuiper, and Watson \eqn{U^2} tests to assess whether residuals
are uniformly distributed on the circle. These tests are commonly used to detect
lack of fit in circular models.
}
\examples{
resids <- runif(100, -pi, pi)
test_residuals_uniformity(resids)

}
